remove FileDownload (because of unknown license)
I was unable to locate/contact the author of this code (Marco Schmidt).
Even though i found other code of him,
which was released to the public domain,
and this code was likely meant to be so too,
we have to assume it is not, as we do not know.
To be on the save side, and because we don't need this file anyway,
i decided to remove it before going GPL.